Added emoticons + updated message bar + added index to messages
