+++ package.json
@@ -19,6 +19,7 @@
     "@pinia/nuxt": "^0.4.7",
     "vite": "^4.2.1",
     "vue": "^3.2.47",
+    "vue3-emoji-picker": "^1.1.7",
     "webpack": "^5.77.0"
   }
 }
